Move GeoTools related value object mappings for jackson from gs-jackson-bindings to gt-jackson-bindings
Refactor mappings from gs-jackson-bindings to gt-jackson-bindings, for
value object types that are GeoTools specific.

New GeoTools filter functions include parameters that weren't used
before, like JTS Point or CoordinateReferenceSystem.

The Jackson mapper for CoordinateReferenceSystem was on the
gs-jackson-bindings module, so take the opportunity to move it to the
GeoTools module alongside other GT value types.
