+++ src/catalog/jackson-bindings/geoserver/src/main/java/org/geoserver/jackson/databind/catalog/dto/Resource.java
@@ -9,6 +9,9 @@ import com.fasterxml.jackson.annotation.JsonSubTypes;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
+import org.geotools.jackson.databind.dto.CRS;
+import org.geotools.jackson.databind.dto.Envelope;
+
 import java.util.List;
 import java.util.Map;
