+++ src/catalog/jackson-bindings/geoserver/src/main/java/org/geoserver/jackson/databind/config/dto/mapper/ConfigInfoMapperConfig.java
@@ -4,10 +4,10 @@
  */
 package org.geoserver.jackson.databind.config.dto.mapper;
 
-import org.geoserver.jackson.databind.catalog.mapper.ValueMappers;
+import org.geoserver.jackson.databind.catalog.mapper.GeoServerValueObjectsMapper;
 import org.geoserver.jackson.databind.mapper.InfoReferenceMapper;
 import org.geoserver.jackson.databind.mapper.PatchMapper;
-import org.geoserver.jackson.databind.mapper.SharedMappers;
+import org.geotools.jackson.databind.filter.mapper.GeoToolsValueMappers;
 import org.mapstruct.MapperConfig;
 import org.mapstruct.ReportingPolicy;
 
@@ -17,8 +17,8 @@ import org.mapstruct.ReportingPolicy;
         uses = {
             ObjectFacotries.class,
             WPSMapper.class,
-            ValueMappers.class,
-            SharedMappers.class,
+            GeoToolsValueMappers.class,
+            GeoServerValueObjectsMapper.class,
             InfoReferenceMapper.class,
             PatchMapper.class
         })
