+++ src/catalog/jackson-bindings/geoserver/src/main/java/org/geoserver/jackson/databind/config/dto/mapper/WPSMapper.java
@@ -6,20 +6,21 @@ package org.geoserver.jackson.databind.config.dto.mapper;
 
 import lombok.Generated;
 
-import org.geoserver.jackson.databind.catalog.mapper.ValueMappers;
+import org.geoserver.jackson.databind.catalog.mapper.GeoServerValueObjectsMapper;
 import org.geoserver.jackson.databind.config.dto.Service;
-import org.geoserver.jackson.databind.mapper.SharedMappers;
+import org.geoserver.jackson.databind.config.dto.Service.WpsService.ProcessGroup;
 import org.geoserver.wps.ProcessGroupInfo;
 import org.geoserver.wps.ProcessGroupInfoImpl;
 import org.geoserver.wps.ProcessInfo;
 import org.geoserver.wps.ProcessInfoImpl;
+import org.geotools.jackson.databind.filter.mapper.GeoToolsValueMappers;
 import org.mapstruct.AnnotateWith;
 import org.mapstruct.Mapper;
 import org.mapstruct.ObjectFactory;
 import org.mapstruct.ReportingPolicy;
 
 @Mapper(
-        uses = {ValueMappers.class, SharedMappers.class},
+        uses = {GeoServerValueObjectsMapper.class, GeoToolsValueMappers.class},
         unmappedTargetPolicy = ReportingPolicy.ERROR)
 @AnnotateWith(value = Generated.class)
 public interface WPSMapper {
@@ -35,4 +36,8 @@ public interface WPSMapper {
     ProcessInfo map(Service.WpsService.Process p);
 
     Service.WpsService.Process map(ProcessInfo p);
+
+    ProcessGroup map(ProcessGroupInfo info);
+
+    ProcessGroupInfo map(ProcessGroup dto);
 }
