+++ src/catalog/jackson-bindings/geotools/src/main/java/org/geotools/jackson/databind/dto/NameDto.java
@@ -2,12 +2,13 @@
  * (c) 2020 Open Source Geospatial Foundation - all rights reserved This code is licensed under the
  * GPL 2.0 license, available at the root application directory.
  */
-package org.geoserver.jackson.databind.catalog.dto;
+package org.geotools.jackson.databind.dto;
 
 import lombok.Data;
 
+/** DTO for {@link org.geotools.api.feature.type.Name} */
 @Data
-public class CRS {
-    private String srs;
-    private String WKT;
+public class NameDto {
+    private String namespaceURI;
+    private String localPart;
 }
