+++ src/catalog/jackson-bindings/geotools/src/main/java/org/geotools/jackson/databind/dto/NumberRangeDto.java
@@ -2,16 +2,14 @@
  * (c) 2020 Open Source Geospatial Foundation - all rights reserved This code is licensed under the
  * GPL 2.0 license, available at the root application directory.
  */
-package org.geoserver.jackson.databind.catalog.dto;
+package org.geotools.jackson.databind.dto;
 
 import lombok.Data;
 
-import org.geoserver.config.util.XStreamPersister;
 import org.geotools.util.NumberRange;
 
 /**
  * @see NumberRange
- * @see XStreamPersister#NumberRangeConverter
  */
 @Data
 public class NumberRangeDto {
