+++ src/catalog/jackson-bindings/geotools/src/main/java/org/geotools/jackson/databind/filter/GeoToolsFilterModule.java
@@ -5,6 +5,8 @@
 package org.geotools.jackson.databind.filter;
 
 import com.fasterxml.jackson.core.Version;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 
@@ -14,18 +16,28 @@ import org.geotools.api.filter.Filter;
 import org.geotools.api.filter.capability.FunctionName;
 import org.geotools.api.filter.expression.Expression;
 import org.geotools.api.filter.sort.SortBy;
+import org.geotools.api.referencing.crs.CoordinateReferenceSystem;
+import org.geotools.api.util.InternationalString;
+import org.geotools.geometry.jts.ReferencedEnvelope;
+import org.geotools.jackson.databind.dto.NameDto;
+import org.geotools.jackson.databind.dto.NumberRangeDto;
+import org.geotools.jackson.databind.dto.VersionDto;
 import org.geotools.jackson.databind.filter.dto.Literal;
 import org.geotools.jackson.databind.filter.dto.LiteralDeserializer;
 import org.geotools.jackson.databind.filter.dto.LiteralSerializer;
 import org.geotools.jackson.databind.filter.mapper.ExpressionMapper;
 import org.geotools.jackson.databind.filter.mapper.FilterMapper;
-import org.geotools.jackson.databind.filter.mapper.ValueMappers;
+import org.geotools.jackson.databind.filter.mapper.GeoToolsValueMappers;
 import org.geotools.jackson.databind.geojson.GeoToolsGeoJsonModule;
 import org.geotools.jackson.databind.util.MapperDeserializer;
 import org.geotools.jackson.databind.util.MapperSerializer;
+import org.geotools.measure.Measure;
+import org.geotools.util.NumberRange;
 import org.locationtech.jts.geom.Geometry;
 import org.mapstruct.factory.Mappers;
 
+import java.util.Locale;
+import java.util.Map;
 import java.util.function.Function;
 
 /**
@@ -64,62 +76,105 @@ public class GeoToolsFilterModule extends SimpleModule {
 
     private static final FilterMapper FILTERS = Mappers.getMapper(FilterMapper.class);
     private static final ExpressionMapper EXPRESSIONS = Mappers.getMapper(ExpressionMapper.class);
-    private static final ValueMappers VALUES = Mappers.getMapper(ValueMappers.class);
+    private static final GeoToolsValueMappers VALUES =
+            Mappers.getMapper(GeoToolsValueMappers.class);
 
     public GeoToolsFilterModule() {
         super(GeoToolsFilterModule.class.getSimpleName(), new Version(1, 0, 0, null, null, null));
 
         log.debug("registering jackson de/serializers for geotools Filter and Expression");
 
-        addSerializer(
+        addMapperSerializer(
                 Expression.class,
-                new MapperSerializer<>(
-                        org.geotools.api.filter.expression.Expression.class, EXPRESSIONS::map));
-        addSerializer(Literal.class, new LiteralSerializer());
-        addDeserializer(
-                Expression.class,
-                new MapperDeserializer<>(
-                        org.geotools.jackson.databind.filter.dto.Expression.class,
-                        EXPRESSIONS::map));
-        addDeserializer(Literal.class, new LiteralDeserializer());
+                EXPRESSIONS::map,
+                org.geotools.jackson.databind.filter.dto.Expression.class,
+                EXPRESSIONS::map);
 
-        addSerializer(Filter.class, new MapperSerializer<>(Filter.class, FILTERS::map));
-        addDeserializer(
+        addSerializer(Literal.class, new LiteralSerializer(), new LiteralDeserializer());
+
+        addMapperSerializer(
                 Filter.class,
-                new MapperDeserializer<>(
-                        org.geotools.jackson.databind.filter.dto.Filter.class, FILTERS::map));
+                FILTERS::map,
+                org.geotools.jackson.databind.filter.dto.Filter.class,
+                FILTERS::map);
 
-        addSerializer(SortBy.class, new MapperSerializer<>(SortBy.class, FILTERS::map));
-        addDeserializer(
+        addMapperSerializer(
                 SortBy.class,
-                new MapperDeserializer<>(
-                        org.geotools.jackson.databind.filter.dto.SortBy.class, FILTERS::map));
+                FILTERS::map,
+                org.geotools.jackson.databind.filter.dto.SortBy.class,
+                FILTERS::map);
 
-        addSerializer(
-                FunctionName.class, new MapperSerializer<>(FunctionName.class, EXPRESSIONS::map));
-        addDeserializer(
+        addMapperSerializer(
                 FunctionName.class,
-                new MapperDeserializer<>(
-                        org.geotools.jackson.databind.filter.dto.Expression.FunctionName.class,
-                        EXPRESSIONS::map));
+                EXPRESSIONS::map,
+                org.geotools.jackson.databind.filter.dto.Expression.FunctionName.class,
+                EXPRESSIONS::map);
 
         addCustomLiteralValueSerializers();
+    }
+
+    private <T> GeoToolsFilterModule addSerializer(
+            Class<T> type, JsonSerializer<T> ser, JsonDeserializer<T> deser) {
+        super.addSerializer(type, ser);
+        super.addDeserializer(type, deser);
+        return this;
     }
 
     /** */
     private void addCustomLiteralValueSerializers() {
         addMapperSerializer(
+                CoordinateReferenceSystem.class,
+                VALUES::crs,
+                org.geotools.jackson.databind.dto.CRS.class,
+                VALUES::crs);
+        addMapperSerializer(
+                CoordinateReferenceSystem.class,
+                VALUES::crs,
+                org.geotools.jackson.databind.dto.CRS.class,
+                VALUES::crs);
+
+        addMapperSerializer(
+                ReferencedEnvelope.class,
+                VALUES::referencedEnvelope,
+                org.geotools.jackson.databind.dto.Envelope.class,
+                VALUES::referencedEnvelope);
+        addMapperSerializer(
                 java.awt.Color.class,
                 VALUES::awtColorToString,
                 String.class,
                 VALUES::stringToAwtColor);
+        addMapperSerializer(
+                org.geotools.api.feature.type.Name.class, VALUES::map, NameDto.class, VALUES::map);
+
+        addMapperSerializer(
+                org.geotools.util.Version.class,
+                VALUES::versionToDto,
+                VersionDto.class,
+                VALUES::dtoToVersion);
+
+        addMapperSerializer(
+                InternationalString.class,
+                VALUES::internationalStringToDto,
+                Map.class,
+                VALUES::dtoToInternationalString);
+
+        addMapperSerializer(
+                NumberRange.class,
+                VALUES::numberRangeToDto,
+                NumberRangeDto.class,
+                VALUES::dtoToNumberRange);
+        addMapperSerializer(
+                Measure.class, VALUES::measureToString, String.class, VALUES::stringToMeasure);
+        addMapperSerializer(
+                Locale.class, VALUES::localeToString, String.class, VALUES::stringToLocale);
     }
 
     /**
      * @param <T> object model type
      * @param <D> DTO type
+     * @return
      */
-    private <T, D> void addMapperSerializer(
+    private <T, D> GeoToolsFilterModule addMapperSerializer(
             Class<T> type,
             Function<T, D> serializerMapper,
             Class<D> dtoType,
@@ -128,7 +183,6 @@ public class GeoToolsFilterModule extends SimpleModule {
         MapperSerializer<T, D> serializer = new MapperSerializer<>(type, serializerMapper);
         MapperDeserializer<D, T> deserializer =
                 new MapperDeserializer<>(dtoType, deserializerMapper);
-        super.addSerializer(type, serializer);
-        super.addDeserializer(type, deserializer);
+        return addSerializer(type, serializer, deserializer);
     }
 }
