+++ src/catalog/jackson-bindings/geotools/src/main/java/org/geotools/jackson/databind/filter/mapper/FilterMapperConfig.java
@@ -13,7 +13,7 @@ import org.mapstruct.ReportingPolicy;
         uses = {
             ExpressionFactory.class,
             FilterFactory.class,
-            ValueMappers.class,
+            GeoToolsValueMappers.class,
             ExpressionMapper.class
         })
 public class FilterMapperConfig {}
